Graph Analysis

ModelGraph Analysis

A ModelGraph supports graph analysis to perform functions such as graph partitioning or community detection. The graph analysis functions are particularly useful for creating decompositions of optimization problems and in fact, this is what is done to use Plasmo's built-in structure-based solvers.

Methods

Plasmo.PlasmoModelGraph.partition — Method.

partition(graph::ModelGraph,n_parts::Int64;alg = :KWAY) –> Vector{Vector{Int64}}

Return a graph partition containing a vector of a vectors of node indices.

source

LightGraphs.label_propagation — Method.

LightGraphs.label_propagation(graph::ModelGraph)

Return partitions corresponding to detected communities using the LightGraphs label propagation algorithm.

source