Modeling

ModelGraph

Constructor

The ModelGraph is the primary object for creating graph-based models in Plasmo.jl. A ModelGraph extends the JuMP.AbstractModel and offers a collection of ModelNodes (which also extend JuMP.AbstractModel) which represent solvable optimization problems. ModelNodes are connected by LinkConstraints over variables which induce underlying LinkEdges. s

A ModelGraph does not require any arguments to construct:

mg = ModelGraph()

A ModelGraph optimizer can be specified in the same way as in JuMP using set_optimizer(::ModelGraph). An optimizer can be any JuMP compatible solver or a custom Plasmo.jl provided solver (see the solvers section). For example, we could construct a ModelGraph that uses the IpoptSolver from the Ipopt package:

graph = ModelGraph()
ipopt = Ipopt.Optimizer
set_optimizer(graph,ipopt)

Adding Nodes

ModelNodes can be added to a ModelGraph using the @node macro. For instance, the below piece of code add the node n1 to the modelgraph mg

@node(mg,n1)

It is also possible to create sets of ModelNodes in a single call to @node like shown in the below code snippet. This example creates a 2x2 grid of modelnodes.

@node(mg,nodes[1:2,1:2])
for node in nodes
    @variable(node,x>=0)
end

We can iterate over the nodes in a ModelGraph using the getnodes function. For example

for node in getnodes(mg)
    println(node)
end

will print the string for every node in the ModelGraph graph.

Variables within a ModelNode can be accessed directly from their enclosing node.

@variable(n1,x >= 0)
println(n1[:x])    #accesses variable x on jump_model

Adding LinkConstraints

LinkConstraints are linear constraints that couple variables across different ModelNodes. The simplest way to add LinkConstraints is to use the @linkconstraint macro. This macro accepts the same input as a JuMP @constraint macro and creates linear constraints over multiple nodes within the same graph.

@variable(nodes[1,1],x >= 0)

@linkconstraint(graph,n1[:x] == nodes[1,1][:x])  #creates a linear constraint between nodes n1 and n2

Subgraph Structures

It is possible to create subgraphs within a ModelGraph object. This is helpful when a user wants to develop to separate systems and link them together within a higher level graph.

sg1 = ModelGraph()
@node(sg1,nsubs1[1:2])
for node in nsub
    @variable(node,y[1:2] >= 0 )
end
@linkconstraint(sg1,nsubs1[:y][1] == nsubs1[:y][2])  #creates a linear constraint between nodes n1 and n2

sg2 = ModelGraph()
@node(sg2,nsubs2[1:2])
for node in nsub
    @variable(node,y[1:2] >= 0 )
end
@linkconstraint(sg2,nsubs2[:y][1] == nsubs2[:y][2])  #creates a linear constraint between nodes n1 and n2

add_subgraph!(mg,sg1)
add_subgraph!(mg,sg2)

@linkconstraint(mg,nsubs1[:y][2]) == nsubs2[:y][2])

Methods

ModelGraph

The ModelGraph contains the following useful methods:

Plasmo.ModelGraph — Type.

ModelGraph()

The ModelGraph Type. Represents a graph containing models (nodes) and the linkconstraints (edges) between them.

source

ModelNode

ModelNodes contain methods for managing their contained JuMP models.

Plasmo.ModelNode
Plasmo.@node(graph::ModelGraph)

Attributes

Plasmo.getnodes
Plasmo.all_nodes
Plasmo.getlinkconstraints
Plasmo.all_linkconstraints