ModelGraph

ModelGraph

Constructor

The ModelGraph is the primary object for creating graph-based models in Plasmo.jl. A ModelGraph is a collection of ModelNodes which are connected by means of LinkingEdges (link constraints) over variables. One way to think of the structure of a ModelGraph is a HyperGraph wherein edges represent linking constraints that can link multiple ModelNode variables.

A ModelGraph does not require any arguments to construct:

mg = ModelGraph()

A ModelGraph solver can be specified upon construction using the solver keyword argument. A solver can be any JuMP compatible solver or a Plasmo provided solver (see solvers section). For example, we could construct a ModelGraph that starts with the IpoptSolver from the Ipopt package:

mg = ModelGraph(solver = IpoptSolver())

Adding Nodes

Nodes can be added to a ModelGraph using the add_node! function. By default, a node contains an empty JuMP Model object.

n1 = add_node!(mg)

A model can be set upon creation by providing a second argument. For example:

model = JuMP.Model()
n1 = add_node!(mg,model)

where model is a JuMP Model object. We can also set a model on a node after construction:

setmodel(n1,model)

This can be helpful in instances where a user wants to swap out a model on a node without changing the graph topology. Keep in mind however that swapping out a model will by default remove any link-constraints that involve that node.

We can also iterate over the nodes in a ModelGraph using the getnodes function. For example

for node in getnodes(mg)
    println(node)
end

will print the string for every node in the ModelGraph mg.

ModelNodes can also be retrieved based on their index, or a node index can be found within a ModelGraph. For example, since n1 was the first node added to mg, it will have an index of 1.

n1 = getnode(mg,1)
getindex(mg,n1) == 1  #will return true

Variables within a JuMP Model can be accessed directly from their enclosing node.

jump_model = Model()
@variable(jump_model,x >= 0)
setmodel(n1,jump_model)
println(n1[:x])  

Adding Link-Constraints

Link constraints are linear constraints that couple variables across different ModelNodes. The simplist way to add link-constraints is to use the @linkconstraint macro. This macro accepts the same input as a JuMP @constraint macro, except it handles linear constraints over multiple nodes within the same graph.

jump_2 = Model()
@variable(jump_2,x >= 0)
n2 = add_node!(mg,jump_2)

@linkconstraint(mg,n1[:x] == n2[:x])

Subgraph Structures

Finally, it is possible to create subgraphs within a ModelGraph object. This is helpful when a user wants to develop to separate systems and link them together within a higher level graph.

Methods

ModelGraph

The ModelGraph contains the following useful methods:

Plasmo.PlasmoModelGraph.ModelGraph — Type.

ModelGraph()

The ModelGraph Type. Represents a graph containing models (nodes) and the linkconstraints (edges) between them. A ModelGraph wraps a BasePlasmoGraph and can use its methods. A ModelGraph also wraps a LinkModel object which extends a JuMP AbstractModel to provide model management functions.

source

JuMP.getobjectivevalue — Function.

Get the ModelGraph objective value

source

Get node objective value

source

Plasmo.PlasmoModelGraph.getinternaljumpmodel — Function.

Get the current created JuMP model for the ModelGraph. Only created when solving using a JuMP compliant solver.

source

JuMP.setsolver — Method.

setsolver(model::AbstractModelGraph,solver::AbstractMathProgSolver)

Set the graph solver to use an AbstractMathProg compliant solver

source

JuMP.setsolver — Method.

setsolver(model::AbstractModelGraph,solver::AbstractPlasmoSolver)

Set the graph solver to use an AbstractMathProg compliant solver

source

Plasmo.PlasmoModelGraph.getsolver — Function.

Get the ModelGraph solver

source

ModelNode

ModelNodes contain methods for managing their contained JuMP models.

Plasmo.PlasmoModelGraph.ModelNode — Type.

The ModelNode type

ModelNode()

Creates an empty ModelNode. Does not add it to a graph.

source

Plasmo.PlasmoGraphBase.add_node! — Method.

add_node!(graph::AbstractModelGraph)

Add a ModelNode to a ModelGraph.

source

Plasmo.PlasmoModelGraph.setmodel — Method.

setmodel(node::ModelNode,m::AbstractModel)

Set the model on a node. This will delete any link-constraints the node is currently part of

source

Plasmo.PlasmoModelGraph.is_nodevar — Function.

is_nodevar(node::ModelNode,var::AbstractJuMPScalar)

Check whether a JuMP variable belongs to a ModelNode

source

Plasmo.PlasmoGraphBase.getnode — Method.

getnode(model::AbstractModel)

Get the ModelNode corresponding to a JuMP Model

source

Plasmo.PlasmoGraphBase.getnode — Method.

getnode(model::AbstractModel)

Get the ModelNode corresponding to a JuMP Variable

source

LinkConstraints

Plasmo.PlasmoModelGraph.getlinkconstraints — Method.

getlinkconstraints(graph::AbstractModelGraph)

Return Array of all LinkConstraints in the ModelGraph graph

source

Plasmo.PlasmoModelGraph.getsimplelinkconstraints — Function.

getsimplelinkconstraints(model::AbstractModelGraph)

Retrieve link-constraints that only connect two nodes"

source

Plasmo.PlasmoModelGraph.gethyperlinkconstraints — Function.

gethyperlinkconstraints(model::AbstractModelGraph)

Retrieve link-constraints that connect three or more nodes"

source

Plasmo.PlasmoModelGraph.get_all_linkconstraints — Function.

getalllinkconstraints(graph::AbstractModelGraph)

Get a list containing every link constraint in the graph, including its subgraphs

source

Plasmo.PlasmoModelGraph.addlinkconstraint — Function.

Add a single link-constraint to the ModelGraph

source

Add a vector of link-constraints to the ModelGraph

source

Plasmo.PlasmoModelGraph.getlinkconstraints — Method.

getlinkconstraints(graph::AbstractModelGraph,node::ModelNode)

Return Array of LinkConstraints for the node

source

Plasmo.PlasmoModelGraph.getlinkconstraints — Method.

getlinkconstraints(node::ModelNode)

Return a Dictionary of LinkConstraints for each graph the node is a member of

source